nix-derivation-parser: switch derivation outputs to hashmap
Derivation outputs are unique by their output path. `nix derivation
show` and the Haskell-Nix-Derivation-Library make use the output as a
key and the rest of the content as a value. Modified the
parse_derivation_output and parse_derivation_outputs functions to mimic
this behavior. Updated the tests to verify it all works.