nix-derivation-parser: update release packages generating scripts
Updated the release-packages.nix file to work for both non-content
addressed and content addressed derivations. It also flattens the
returning attrset which might be necessary as nix-instantiate was not
instantiating the ca-derivations unless the attrset was 1 level deep.

Modified the original script to work properly. I am fairly certain that
nix-instantiate was not actually instantiating derivations with the
`--eval` flag. So now it works in two stages: the first instantiates the
derivations into the nix store and the second pipes the same output to
json for processing.

Added a second script to generate the ca-derivation files.

Updated the .gitignore to ignore all the newly added derivations.