meta: add pre-commit hooks via git-hooks.nix
Adds `git-hooks.nix` as a flake input. Adds `flake-compat` and
`gitignore.nix` as flake inputs as well so we can pin all the inputs of
`git-hooks.nix`. Set all the recently added inputs to follow inputs we
specify.

Added `pre-commit` checks and used the `treefmt` option with the
`treefmt` package defined by `treefmt-nix`. Added packages and a shell
hook to the `bootstrap` devShell so it will install pre-commit
configuration.

Added `.pre-commit-config.yaml` to the .gitignore as instructed.